Refactored testingsetup
