+++ cypress/support/config.js
@@ -1,0 +1,17 @@
+export default {
+  baseUrl: 'http://localhost:3000',
+  users: {
+    admin: {
+      email: 'admin@example.org',
+      password: 1234
+    },
+    moderator: {
+      email: 'moderator@example.org',
+      password: 1234
+    },
+    user: {
+      email: 'user@example.org',
+      password: 1234
+    }
+  }
+}
